* Added a singleCachedStateStrategy * Moved all strategies into a package
